Moving all logic into waf, removing _waf. Removed engine flip option, cleanup code
