+++ include/waflz/waf.h
@@ -25,8 +25,9 @@
 //: ----------------------------------------------------------------------------
 //: includes
 //: ----------------------------------------------------------------------------
-#include <waflz/_waf.h>
 #include "waflz/parser.h"
+#include <waflz/config_parser.h>
+#include "waflz/def.h"
 #include <set>
 //: ----------------------------------------------------------------------------
 //: constants
@@ -50,9 +51,13 @@ namespace ns_waflz {
 class rqst_ctx;
 class ac;
 class byte_range;
+class engine;
+class regex;
+class profile;
 //: ----------------------------------------------------------------------------
 //: types
 //: ----------------------------------------------------------------------------
+typedef std::list<regex *> pcre_list_t;
 typedef std::list<regex *> regex_list_t;
 typedef std::list<ac *> ac_list_t;
 typedef std::list<byte_range *> byte_range_list_t;
@@ -75,7 +80,7 @@ typedef struct _compiled_config {
 //: ----------------------------------------------------------------------------
 //: \details: TODO
 //: ----------------------------------------------------------------------------
-class waf: public _waf
+class waf
 {
 public:
         // -------------------------------------------------
@@ -86,7 +91,24 @@ public:
         int32_t process(waflz_pb::event **ao_event, void *a_ctx);
         int32_t init(profile &a_profile, bool a_leave_tmp_file = false);
         int32_t init(config_parser::format_t a_format, const std::string &a_path, bool a_apply_defaults = false);
+        int32_t regex_list_add(const std::string &a_regex, pcre_list_t &a_pcre_list);
         int32_t get_str(std::string &ao_str, config_parser::format_t a_format);
+        const char *get_err_msg(void) { return m_err_msg; }
+        // -------------------------------------------------
+        // properties
+        // -------------------------------------------------
+        void set_id(const std::string &a_id) { m_id = a_id; }
+        void set_name(const std::string &a_name) { m_name = a_name; }
+        void set_owasp_ruleset_version(uint32_t a_version) { m_owasp_ruleset_version = a_version; }
+        uint32_t get_owasp_ruleset_version(void) { return m_owasp_ruleset_version; }
+        // -------------------------------------------------
+        // public static methods
+        // -------------------------------------------------
+        static int32_t append_rqst_info(waflz_pb::event &ao_event, void *a_ctx);
+        // TODO -make private!!!
+        pcre_list_t m_il_query;
+        pcre_list_t m_il_header;
+        pcre_list_t m_il_cookie;
 private:
         // -------------------------------------------------
         // private types
@@ -124,6 +146,16 @@ private:
         directive_list_t m_mx_rule_list;
 #ifdef WAFLZ_NATIVE_ANOMALY_MODE
         int32_t m_anomaly_score_cur;
+        bool m_is_initd;
+        char m_err_msg[WAFLZ_ERR_LEN];
+        engine &m_engine;
+        // -------------------------------------------------
+        // properties
+        // -------------------------------------------------
+        std::string m_id;
+        std::string m_name;
+        uint32_t m_owasp_ruleset_version;
+        bool m_no_log_matched;
 #endif
         // -------------------------------------------------
         // sharing private fields with engine...
