+++ src/CMakeLists.txt
@@ -139,7 +139,6 @@ SET(SRCS
     # --------------------------------------------
     # core
     # --------------------------------------------
-    core/_waf.cc
     core/waf.cc
     core/acl.cc
     core/engine.cc
